Moved more code from global helpers into delegations helpers
